requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
pydicom>=2.3.0
nibabel>=4.0.0
matplotlib>=3.5.0
scikit-image>=0.19.0
scipy>=1.7.0
torch>=1.11.0
torchvision>=0.12.0
monai>=1.0.0
PySide6>=6.3.0
reportlab>=3.6.0
Pillow>=9.0.0
tqdm>=4.64.0
SimpleITK>=2.2.0
vtk>=9.1.0
pandas>=1.4.0
seaborn>=0.11.0 